Create show_batch.py
This function is for displaying multiple randomly selected images from an image dataset. As arguments, it receives the image dataset (img_ds), the number of images to display (num_items), the number of rows and columns of images to display (num_rows, num_cols), and the prediction array (predict_arr). predict_arr is used to hold information such as the predicted class for the image, but this argument is optional.